feat: Add Knowledge Graphic

+++ src/main/java/top/rslly/iot/services/thingsModel/ProductServiceImpl.java
@@ -70,6 +70,12 @@ public class ProductServiceImpl implements ProductService {
   private ProductVoiceDiyRepository productVoiceDiyRepository;
   @Resource
   private AdminConfigRepository adminConfigRepository;
+  @Resource
+  private KnowledgeGraphicNodeRepository knowledgeGraphicNodeRepository;
+  @Resource
+  private KnowledgeGraphicRelationRepository knowledgeGraphicRelationRepository;
+  @Resource
+  private KnowledgeGraphicAttributeRepository knowledgeGraphicAttributeRepository;
 
 
   @Override
@@ -191,6 +197,8 @@ public class ProductServiceImpl implements ProductService {
         productVoiceDiyRepository.findAllByProductId(id);
     List<AdminConfigEntity> adminConfigEntityList =
         adminConfigRepository.findAllBySetKey("wx_default_product");
+    List<KnowledgeGraphicNodeEntity> knowledgeGraphicNodeList =
+            knowledgeGraphicNodeRepository.findAllByProductUid(id);
     boolean p1 = productModelEntityList.isEmpty();
     boolean p2 = wxProductBindEntityList.isEmpty();
     boolean p3 = otaEntityList.isEmpty();
@@ -201,7 +209,8 @@ public class ProductServiceImpl implements ProductService {
     boolean p8 = productToolsBanEntityList.isEmpty();
     boolean p9 = agentLongMemoryEntityList.isEmpty();
     boolean p10 = productVoiceDiyEntityList.isEmpty();
-    if (p1 && p2 && p3 && p4 && p5 && p6 && p7 && p8 && p9 && p10) {
+    boolean p11 = knowledgeGraphicNodeList.isEmpty();
+    if (p1 && p2 && p3 && p4 && p5 && p6 && p7 && p8 && p9 && p10 && p11) {
       List<ProductEntity> result = productRepository.deleteById(id);
       if (result.isEmpty())
         return ResultTool.fail(ResultCode.PARAM_NOT_VALID);
@@ -222,6 +231,14 @@ public class ProductServiceImpl implements ProductService {
             }
           } catch (Exception ignore) {
           }
+        }
+        if(!knowledgeGraphicNodeList.isEmpty()){
+            for(KnowledgeGraphicNodeEntity knowledgeGraphicNodeEntity : knowledgeGraphicNodeList){
+                long kId = knowledgeGraphicNodeEntity.getId();
+                knowledgeGraphicRelationRepository.deleteAllByFrom(kId);
+                knowledgeGraphicAttributeRepository.deleteByBelong(kId);
+            }
+            knowledgeGraphicNodeRepository.deleteAllByProductUid(id);
         }
         return ResultTool.success(result);
       }
